Update README.md
I'm thrilled to share my recent cricket analytics project journey. In Jupyter Notebook, I harnessed the power of Python and Pandas to clean and refine the cricket data, ensuring accuracy and clarity.

To enhance organization, Excel stepped in to neatly structure the numbers and create a seamless flow.

Finally, the data was brought to life with a visually appealing cricket dashboard using Power BI.

If you're interested in the technical details: Python and Pandas in Jupyter Notebook for data cleaning, Excel for organization, and Power BI for dynamic visualization.